requirements.txt
aiogram==2.20
aioredis==2.0.1
gino==1.0.1
python-dotenv==0.20.0
SQLAlchemy==1.3.24
requests==2.27.1